+++ packages/terra-data-grid/CHANGELOG.md
@@ -7,6 +7,7 @@
 
 * Added
   * Added `hasVisibleColumnHeaders` prop for FlowsheetDataGrid to toggle visibility of column headers.
+  * Added section support to FlowsheetDataGrid.
 
 Changed
   * Updated all columns and cells to be selectable in FlowsheetDataGrid, as `isSelectable` prop is not supported for columns or cells.